scripts/requirements.txt
psycopg2-binary
requests
playwright
faker
